+++ source/geh_calculated_measurements/tests/subsystem_tests/environment_configuration.py
@@ -1,10 +1,8 @@
-from pathlib import Path
-
 from geh_common.application.settings import ApplicationSettings
 from pydantic import Field
 from pydantic_settings import SettingsConfigDict
 
-PROJECT_ROOT = Path(__file__).parent.parent.parent
+from tests import TESTS_ROOT
 
 
 class EnvironmentConfiguration(ApplicationSettings):
@@ -24,6 +22,6 @@ class EnvironmentConfiguration(ApplicationSettings):
     catalog_name: str = Field(alias="SHARED_CATALOG_NAME")
 
     model_config = SettingsConfigDict(
-        env_file=f"{PROJECT_ROOT}/.env",
+        env_file=f"{TESTS_ROOT}/.env",
         env_file_encoding="utf-8",
     )